// ts/react/common/types/bedev2Types.ts
export enum TContentType {
	Game = "Game",
	CatalogAsset = "CatalogAsset",
	CatalogBundle = "CatalogBundle",
}

export enum TTreatmentType {
	Carousel = "Carousel",
	AvatarCarousel = "AvatarCarousel",
	SortlessGrid = "SortlessGrid",
	FriendCarousel = "FriendCarousel",
	InterestGrid = "InterestGrid",
	Pills = "Pills",
	Sdui = "sdui",
	SongCarousel = "SongCarousel",
}
// ...
export type TOmniRecommendationGame = {
	contentType: TContentType.Game;
	contentId: number;
	contentMetadata: Record<string, string>;
};
export type TOmniRecommendationCatalog = {
	contentType: TContentType.CatalogAsset | TContentType.CatalogBundle;
	contentId: number;
};
export type TOmniRecommendation =
	| TOmniRecommendationGame
	| TOmniRecommendationCatalog;

export type TTopicLayoutData = {
	componentType?: TComponentType;
	playerCountStyle?: TPlayerCountStyle;
	playButtonStyle?: TPlayButtonStyle;
	hoverStyle?: THoverStyle;
	infoText?: string;
	hideSeeAll?: "true" | "false";
	navigationRootPlaceId?: string;
	isSponsoredFooterAllowed?: "true" | "false";
	linkPath?: string;
	subtitleLinkPath?: string;
	endTimestamp?: string;
	countdownString?: string;
	backgroundImageAssetId?: string;
};

type TSharedGameSort = {
	topic: string;
	topicId: number;
	treatmentType:
		| TTreatmentType.Carousel
		| TTreatmentType.SortlessGrid
		| TTreatmentType.InterestGrid;
	subtitle?: string;
	topicLayoutData?: TTopicLayoutData;
};

export type TOmniRecommendationSduiSort = {
	topicId: number;
	treatmentType: TTreatmentType.Sdui;
	numberOfRows?: number;

	feedItemKey: string;
};

export type TOmniRecommendationGameSort = TSharedGameSort & {
	recommendationList: TOmniRecommendationGame[];
	numberOfRows?: number;
};

export type TOmniRecommendationCatalogSort = {
	topic: string;
	topicId: number;
	treatmentType: TTreatmentType.AvatarCarousel;
	recommendationList: TOmniRecommendationCatalog[];
	numberOfRows?: number;
};

export type TOmniRecommendationFriendSort = {
	topic: string;
	topicId: number;
	treatmentType: TTreatmentType.FriendCarousel;
	numberOfRows?: number;
};

export type TFilterOption = {
	optionId: string;
	optionDisplayName: string;
};

export type TFiltersData = {
	filterId: string;
	filterDisplayName: string;
	filterType: string;
	filterOptions: TFilterOption[];
	selectedOptionId: string;
};
// ...
export type TExploreApiGameSort = TSharedGameSort & {
	games: TGameData[];
	sortId: string;
	contentType: string;
	nextPageToken: string;
	gameSetTargetId?: number;
	appliedFilters?: string;
};

export type TExploreApiFiltersSort = {
	topic: string;
	topicId: number;
	treatmentType: TTreatmentType.Pills;
	subtitle?: string;
	topicLayoutData?: TTopicLayoutData;
	filters: TFiltersData[];
	nextPageToken: string;
	sortId: string;
	contentType: string;
	gameSetTargetId?: number;
};

export type TExploreApiSongsSort = {
	topic: string;
	topicId: number;
	songs: TSongData[];
	treatmentType: TTreatmentType.SongCarousel;
	nextPageToken: string;
	sortId: string;
	contentType: string;
	subtitle?: string;
	topicLayoutData?: TTopicLayoutData;
};

export type TSongSort = TExploreApiSongsSort;

export type TExploreApiSort =
	| TExploreApiGameSort
	| TExploreApiFiltersSort
	| TExploreApiSongsSort;

export type TExploreApiSorts = {
	sorts: TExploreApiSort[];
	nextSortsPageToken: string;
};

export type TGameSort = TExploreApiGameSort | TOmniRecommendationGameSort;

export type TOmniRecommendationSort =
	| TOmniRecommendationGameSort
	| TOmniRecommendationCatalogSort
	| TOmniRecommendationFriendSort
	| TOmniRecommendationSduiSort;

export type TSort =
	| TGameSort
	| TOmniRecommendationCatalogSort
	| TOmniRecommendationFriendSort
	| TExploreApiFiltersSort
	| TOmniRecommendationSduiSort
	| TSongSort;